made test timeout manipulation opt-in via adjustTestsTimeouts()
