+++ packages/testing/src/timeouts.ts
@@ -1,12 +1,47 @@
-import { isDebugMode } from './debug-tests';
 import { mochaCtx } from './mocha-ctx';
+
+const getDebug = () => {
+    const debug = (globalThis as { process?: { env: { DEBUG?: string } } })?.process?.env?.DEBUG;
+    return debug === 'true' || parseInt(debug || '0') > 0;
+};
+
+if (getDebug()) {
+    // eslint-disable-next-line no-console
+    console.log('Testing in debug mode');
+}
+
+const forcedDebugModes = new Map<Mocha.Context, boolean>();
+
+/**
+ *
+ * @returns true if DEBUG=positive int/true or mocha is running in debug mode
+ */
+export function isDebugMode() {
+    const ctx = mochaCtx();
+    if (ctx && forcedDebugModes.has(ctx)) {
+        return forcedDebugModes.get(ctx);
+    }
+    return getDebug() || mochaCtx()?.timeout() === 0;
+}
+
+/**
+ * override the DEBUG environment variable for the current test
+ * @param value
+ */
+export function overrideDebugMode(value: boolean) {
+    const ctx = mochaCtx();
+    if (!ctx) {
+        throw new Error('No mocha context');
+    }
+    forcedDebugModes.set(ctx, value);
+}
 
 const forcedTimeoutScale = new Map<Mocha.Context, number>();
 /**
- * 
+ *
  * @returns the TIMEOUT_SCALE environment variable, or 1 if not set
  */
-export function getTimeoutScale  ()  {
+export function getTimeoutScale() {
     const ctx = mochaCtx();
     if (ctx && forcedTimeoutScale.has(ctx)) {
         return forcedTimeoutScale.get(ctx)!;
@@ -22,7 +57,7 @@ export function getTimeoutScale  ()  {
     }
 
     return multiplier;
-};
+}
 
 if (getTimeoutScale() !== 1) {
     // eslint-disable-next-line no-console
@@ -72,10 +107,14 @@ export function locatorTimeout(ms = 10_000) {
 }
 
 /**
- * Manipulates the default timeouts for tests
+ * Adjust tests timeouts based on DEBUG and TIMEOUT_SCALE environment variables
  */
-if (!isDebugMode() && getTimeoutScale() !== 1) {
-    beforeEach('wrap mocha runnables to save ctx', function () {
-        this.currentTest?.timeout(this.currentTest?.timeout() *  getTimeoutScale());
+export function adjustTestsTimeouts() {
+    beforeEach(`adjust test timeout to env: DEBUG & TIMEOUT_SCALE `, function () {
+        if (isDebugMode()) {
+            this.timeout(0);
+        } else {
+            this.timeout(this.timeout() * getTimeoutScale());
+        }
     });
 }
